Reorganized the udfs directories
