+++ exasol_transformers_extension/udfs/callers/filling_mask_udf_call.py
@@ -1,0 +1,8 @@
+from exasol_transformers_extension.udfs.models.filling_mask_udf import FillingMask
+
+
+udf = FillingMask(exa)
+
+
+def run(ctx):
+    return udf.run(ctx)
