+++ exasol_transformers_extension/udfs/callers/question_answering_udf_call.py
@@ -1,4 +1,4 @@
-from exasol_transformers_extension.udfs.question_answering_udf \
+from exasol_transformers_extension.udfs.models.question_answering_udf \
     import QuestionAnswering
 
 udf = QuestionAnswering(exa)
