+++ exasol_transformers_extension/udfs/callers/sequence_classification_single_text_udf_call.py
@@ -1,4 +1,4 @@
-from exasol_transformers_extension.udfs.sequence_classification_single_text_udf \
+from exasol_transformers_extension.udfs.models.sequence_classification_single_text_udf \
     import SequenceClassificationSingleText
 
 udf = SequenceClassificationSingleText(exa)
