+++ tests/integration_tests/without_db/test_question_answering_udf.py
@@ -6,7 +6,7 @@ import torch
 
 from tests.utils.parameters import model_params
 from exasol_udf_mock_python.connection import Connection
-from exasol_transformers_extension.udfs.question_answering_udf import \
+from exasol_transformers_extension.udfs.models.question_answering_udf import \
     QuestionAnswering
