+++ tests/unit_tests/udf_wrapper_params/sequence_classification/multiple_bfsconn_single_subdir_single_model_multiple_batch.py
@@ -7,8 +7,7 @@ from tests.unit_tests.udf_wrapper_params.sequence_classification.\
 
 def udf_wrapper_single_text():
     from exasol_udf_mock_python.udf_context import UDFContext
-    from exasol_transformers_extension.udfs. \
-        sequence_classification_single_text_udf import \
+    from exasol_transformers_extension.udfs.models.sequence_classification_single_text_udf import \
         SequenceClassificationSingleText
     from tests.unit_tests.udf_wrapper_params.sequence_classification. \
         mock_sequence_tokenizer import MockSequenceTokenizer
@@ -28,8 +27,7 @@ def udf_wrapper_single_text():
 
 def udf_wrapper_text_pair():
     from exasol_udf_mock_python.udf_context import UDFContext
-    from exasol_transformers_extension.udfs.\
-        sequence_classification_text_pair_udf import \
+    from exasol_transformers_extension.udfs.models.sequence_classification_text_pair_udf import \
         SequenceClassificationTextPair
     from tests.unit_tests.udf_wrapper_params.sequence_classification. \
         mock_sequence_tokenizer import MockSequenceTokenizer
